Refactor enforce_distribution and update test expectations
Add necessary imports to enforce_distribution.rs and remove unused
BoundedWindowAggExec import. Move the test for superset hash keys
to a more logical position, ensuring consistency in testing for
hash join partitioning.

Update test expectations to reflect current optimizer behavior
by removing unnecessary superset repartitions and replacing
them with exact match requirements. Revise comments and
assertions to align with the improved optimization.
Delete the original enforce_distribution_superset.rs file.
